fix issue with presentation display and create an unscheduled tab for presentations.

+++ supabase/migrations/20260312000000_add_buffer_to_presentations.sql
@@ -1,1 +1,2 @@
-alter table public.presentations add column if not exists buffer integer;
+alter table public.presentations
+    add column if not exists buffer smallint;
